Put classes into package
