+++ data/dazAI.txt
@@ -1,5 +1,4 @@
 D | 0 | return book | 2025-02-01 1800
-D | 0 | return book | 2025-02-01 1800
-E | 0 | project meeting | 2025-02-02 1400 | 2025-02-02 1600
 T | 0 | borrow book
 D | 0 | return book | 2025-09-30 1800
+T | 0 | do assignment
