+++ src/main/java/dazai/AddCommand.java
@@ -1,3 +1,5 @@
+package dazai;
+
 import java.io.IOException;
 
 public class AddCommand extends Command {
