+++ src/main/java/dazai/Command.java
@@ -1,3 +1,5 @@
+package dazai;
+
 public abstract class Command {
     public abstract void execute(TaskList tasks, Ui ui, Storage storage) throws DazAIException;
