+++ src/main/java/dazai/DazAI.java
@@ -1,3 +1,5 @@
+package dazai;
+
 import java.io.IOException;
 
 public class DazAI {
