+++ src/main/java/dazai/DazAIException.java
@@ -1,3 +1,5 @@
+package dazai;
+
 public class DazAIException extends Exception{
     public DazAIException(String message) {
         super(message);
