+++ src/main/java/dazai/Deadline.java
@@ -1,3 +1,5 @@
+package dazai;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
