+++ src/main/java/dazai/Event.java
@@ -1,3 +1,5 @@
+package dazai;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
