+++ src/main/java/dazai/ExitCommand.java
@@ -1,3 +1,5 @@
+package dazai;
+
 public class ExitCommand extends Command {
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) {
