+++ src/main/java/dazai/Task.java
@@ -1,3 +1,5 @@
+package dazai;
+
 public class Task {
 
     protected String description;
