+++ src/main/java/dazai/TaskList.java
@@ -1,3 +1,5 @@
+package dazai;
+
 import java.util.ArrayList;
 
 public class TaskList {
