+++ src/main/java/dazai/Ui.java
@@ -1,3 +1,5 @@
+package dazai;
+
 import java.util.ArrayList;
 import java.util.Scanner;
